gvc-channel-bar: Fix -Wbad-function-cast warning

+++ mate-volume-control/gvc-channel-bar.c
@@ -381,9 +381,9 @@ static gboolean on_scale_button_press_event(GtkWidget *widget,
   if (bar->priv->control_flags & MATE_MIXER_STREAM_CONTROL_MUTE_READABLE &&
       bar->priv->control_flags & MATE_MIXER_STREAM_CONTROL_VOLUME_READABLE) {
     if (mate_mixer_stream_control_get_mute(bar->priv->control) == TRUE) {
-      guint minimum = (guint)gtk_adjustment_get_lower(bar->priv->adjustment);
+      gdouble minimum = gtk_adjustment_get_lower(bar->priv->adjustment);
 
-      if (mate_mixer_stream_control_get_volume(bar->priv->control) > minimum)
+      if (mate_mixer_stream_control_get_volume(bar->priv->control) > (guint) minimum)
         bar->priv->click_lock = TRUE;
     }
   }
